Added files from first 2 units
These are the first 2 units of my Intro to Programming class. The first unit programs are simple, asking for basic programs to allow for user input, printing to the screen, and evaluating arithmetic expressions. Unit 2 includes a payroll calculator that calculates gross and net pay and displays the working tax rate and a cab fare calculator based on various values given in the assignment.